api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
decorator==5.1.1
dill==0.3.6
Django==4.2
djlint==1.23.3
docutils==0.19
EditorConfig==0.12.3
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
flake8==6.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.4
ipython==8.10.0
isort==5.12.0
jedi==0.18.2
jsbeautifier==1.14.7
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
proto-plus==1.22.3
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==2.5.1
pydantic_core==2.14.3
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.16.2
pyparsing==3.1.1
pyserial==3.5
PyYAML==6.0
regex==2023.3.23
requests==2.30.0
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
thonny==4.0.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.8.0
urllib3==2.0.2
uvicorn==0.24.0.post1
wcwidth==0.2.6
wrapt==1.14.1
